requirements.txt
streamlit>=1.20.0
openai>=1.0.0
pdfplumber>=0.6.0
python-docx>=0.8.11
pytesseract>=0.3.10
pillow>=9.0.0
chromadb>=0.3.24
tqdm>=4.64.0
pandas>=1.4.0
scikit-learn>=1.0.2
python-multipart>=0.0.5
numpy>=1.23.0
